Refactor setup script
Refactor setup script to correctly handle references to local packages
from overlays configuration.

+++ setup.sh
@@ -2,15 +2,25 @@
 
 set -e
 
-# Backup existing configuration
+# Backup existing configuration.
 if [ -d /etc/nixos ]; then
-    cp -r /etc/nixos /etc/nixos.orig
+    mv /etc/nixos /etc/nixos.old
 fi
+rm -rf /etc/nixos && mkdir /etc/nixos
 
-# Cleanup existing configuration
-rm -f /etc/nixos/*
-
-# Create symlinks to new configuration
+# Create symlinks to new configuration.
 for file in "${PWD}"/config/*; do
-    ln -sf "${file}" /etc/nixos/
+    if [ -f "${file}" ]; then
+        # If this is a file then create symlink to it.
+        ln -sf "${file}" /etc/nixos/
+    else
+        # If this is a directory create symlinks to files in it.
+        # We can't symlink directory itself because Nix tools don't handle
+        # relative paths to parent directory correctly in this case.
+        dirname=$(basename "${file}")
+        mkdir /etc/nixos/"${dirname}"
+        for nested in "${file}"/*; do
+            ln -sf "${nested}" /etc/nixos/"${dirname}"
+        done
+    fi
 done
